[buf-list] add a Cursor type
This type is very similar to std::io::Cursor, except it wraps a BufList.

We have to add some version gating to support older Rust versions, which
implies expanding CI, and also checking in a Cargo.lock.

We have to pin tokio to an older version so that we can test it against
Rust 1.51. We choose 1.6.1 because it is the first version to have
AsyncWriteExt::write_all_buf, used in one of the tests.